Fix crash in H263 encoder on 64 bit platforms
The HTFMPrepareCurMB function initialises a 256 byte buffer in the
VideoEncData structure.  This buffer, VideoEncData.currYMB, is
initialised by a for loop that attempts to write 4 32 bit words 16 times.
The problem is that it writes these words using a Long* and so on 64 bit
platforms the code ends up writing 512 bytes and not 256.  The result
is that other VideoEncData members get overwritten and the encoder crashes
later on in EncodeFrameCombinedMode.

+++ media/libstagefright/codecs/m4v_h263/enc/src/motion_est.cpp
@@ -674,11 +674,11 @@ void CalcThreshold(double pf, double exp_lamda[], Int nrmlz_th[])
 void    HTFMPrepareCurMB(VideoEncData *video, HTFM_Stat *htfm_stat, UChar *cur)
 {
     void* tmp = (void*)(video->currYMB);
-    ULong *htfmMB = (ULong*)tmp;
+    uint32_t *htfmMB = (uint32_t*)tmp;
     UChar *ptr, byte;
     Int *offset;
     Int i;
-    ULong word;
+    uint32_t word;
     Int width = video->currVop->width;
 
     if (((Int)video->numVopsInGOP) % 30 == 1)
